Replace the REBUILD_MANUAL option with a path
Instead of the somewhat iffy REBUILD_MANUAL option, there's now a
different HTML_DOCS_DIR setting defaulting to the bundled copy. The
CMakeLists file in doc/src is now standalone. It can be used to build
the HTML docs in a separate build directory, then the main app build can
be pointed at that build directory. This is slightly less convenient,
but the dependencies are much easier to get right.